Use floating point PCM for a2 speaker

+++ clem_types.h
@@ -158,7 +158,7 @@ struct ClemensDeviceAudio {
   bool a2_speaker_tense; /**< the a2 speaker state tense, relax */
   int32_t a2_speaker_frame_count;
   int32_t a2_speaker_frame_threshold;
-  int16_t a2_speaker_level;
+  float a2_speaker_level;
 
   /* host supplied mix buffer */
   struct ClemensAudioMixBuffer mix_buffer;
